refactor: AI Panel 타입 파일명 명확화
